requirements.txt
Django==4.2.7
mysqlclient==2.2.0
celery==5.3.4
redis==5.0.1
requests==2.31.0
django-cors-headers==4.3.0
django-rest-framework==3.14.0
Pillow==10.1.0
python-dotenv==1.0.0
uwsgi==2.0.22
django-celery-beat==2.5.0
django-redis==5.4.0 